# Static strategy instances