--- a/src/GraphQL.ApiTests/GraphQL.SystemReactive.approved.txt
+++ b/src/GraphQL.ApiTests/GraphQL.SystemReactive.approved.txt
@@ -3,6 +3,7 @@ namespace GraphQL.Execution
     public class SubscriptionExecutionStrategy : GraphQL.Execution.ParallelExecutionStrategy
     {
         public SubscriptionExecutionStrategy() { }
+        public static GraphQL.Execution.SubscriptionExecutionStrategy Instance { get; }
         public override System.Threading.Tasks.Task<GraphQL.ExecutionResult> ExecuteAsync(GraphQL.Execution.ExecutionContext context) { }
         protected virtual System.Threading.Tasks.Task<System.IObservable<GraphQL.ExecutionResult>> ResolveEventStreamAsync(GraphQL.Execution.ExecutionContext context, GraphQL.Execution.ExecutionNode node) { }
     }