--- a/src/GraphQL/Execution/ParallelExecutionStrategy.cs
+++ b/src/GraphQL/Execution/ParallelExecutionStrategy.cs
@@ -4,9 +4,14 @@ using GraphQL.DataLoader;
 
 namespace GraphQL.Execution
 {
-    /// <inheritdoc cref="ParallelExecutionStrategy.ExecuteNodeTreeAsync(ExecutionContext, ObjectExecutionNode)"/>
+    /// <inheritdoc cref="ExecuteNodeTreeAsync(ExecutionContext, ObjectExecutionNode)"/>
     public class ParallelExecutionStrategy : ExecutionStrategy
     {
+        /// <summary>
+        /// Gets a static instance of <see cref="ParallelExecutionStrategy"/> strategy.
+        /// </summary>
+        public static ParallelExecutionStrategy Instance { get; } = new ParallelExecutionStrategy();
+
         /// <summary>
         /// Executes document nodes in parallel. Field resolvers must be designed for multi-threaded use.
         /// Nodes that return a <see cref="IDataLoaderResult"/> will execute once all other pending nodes