--- a/src/GraphQL/Execution/SerialExecutionStrategy.cs
+++ b/src/GraphQL/Execution/SerialExecutionStrategy.cs
@@ -4,9 +4,14 @@ using GraphQL.DataLoader;
 
 namespace GraphQL.Execution
 {
-    /// <inheritdoc cref="SerialExecutionStrategy.ExecuteNodeTreeAsync(ExecutionContext, ObjectExecutionNode)"/>
+    /// <inheritdoc cref="ExecuteNodeTreeAsync(ExecutionContext, ObjectExecutionNode)"/>
     public class SerialExecutionStrategy : ExecutionStrategy
     {
+        /// <summary>
+        /// Gets a static instance of <see cref="SerialExecutionStrategy"/> strategy.
+        /// </summary>
+        public static SerialExecutionStrategy Instance { get; } = new SerialExecutionStrategy();
+
         /// <summary>
         /// Executes document nodes serially. Nodes that return a <see cref="IDataLoaderResult"/> will
         /// execute once all other pending nodes have been completed.