``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.14393.3986 (1607/AnniversaryUpdate/Redstone1)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
Frequency=3515611 Hz, Resolution=284.4456 ns, Timer=TSC
.NET Core SDK=5.0.102
  [Host]     : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT
  DefaultJob : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT


```
before
|        Method | UseCaching |         Mean |      Error |     StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-------------:|-----------:|-----------:|---------:|--------:|------:|----------:|
| **Introspection** |      **False** | **1,123.941 μs** | **12.4247 μs** | **11.0142 μs** | **105.4688** | **35.1563** |     **-** | **487.11 KB** |
|          Hero |      False |    18.458 μs |  0.2305 μs |  0.2156 μs |   2.1362 |       - |     - |   8.73 KB |
| **Introspection** |       **True** |   **851.471 μs** | **11.1092 μs** | **10.3915 μs** | **106.4453** | **35.1563** |     **-** | **447.43 KB** |
|          Hero |       True |     3.730 μs |  0.0679 μs |  0.0808 μs |   0.9918 |       - |     - |   4.05 KB |

after
|        Method | UseCaching |         Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-------------:|----------:|----------:|---------:|--------:|------:|----------:|
| **Introspection** |      **False** | **1,112.457 μs** | **7.9694 μs** | **6.6548 μs** | **107.4219** | **33.2031** |     **-** | **487.09 KB** |
|          Hero |      False |    19.294 μs | 0.1714 μs | 0.1520 μs |   2.1057 |       - |     - |    8.7 KB |
| **Introspection** |       **True** |   **842.875 μs** | **6.8388 μs** | **5.7107 μs** | **106.4453** | **35.1563** |     **-** | **447.41 KB** |
|          Hero |       True |     3.722 μs | 0.0520 μs | 0.0487 μs |   0.9842 |       - |     - |   4.03 KB |


I don't know why the time is much longer in the second line, this is some kind of fluctuation on my machine.